test(api-core): move machine creator tests to site explorer
Refactor machine creator integration tests out of api-core and into
the site-explorer crate, preserving coverage while using the shared
test harness.  Keep the DPU discovery state progression coverage in
api-core machine state tests.
