+++ crates/api-core/src/tests/mod.rs
@@ -65,7 +65,6 @@ mod mac_address_pool;
 mod machine_admin_force_delete;
 mod machine_bmc_metadata;
 mod machine_boot_override;
-mod machine_creator;
 mod machine_dhcp;
 mod machine_discovery;
 mod machine_find;
